requirements.txt
APScheduler==3.7.0
certifi==2023.7.22
click==7.1.2
fastapi==0.65.2
h11==0.12.0
kafka-python==2.0.2
orjson==3.5.2
psycopg2-binary==2.8.6
pydantic==1.8.2
redis==3.5.3
sentry-sdk==1.1.0
starlette==0.14.2
typing-extensions==3.10.0.0
urllib3==1.26.5
uvicorn==0.13.4
requests==2.26.0
livekit-server-sdk-python==1.0.0